test: split up node tests
